+++ sender/flight_state.c
@@ -1,7 +1,7 @@
 #include "flight_state.h"
 
-int throttle_pct = 0;
-int bank_angle   = 0;
-int pitch_angle  = 0;
+double throttle_pct = 0;
+double bank_angle   = 0;
+double pitch_angle  = 0;
 double altitude  = 0;
-int rpm          = 0;
+double rpm          = 0;
