+++ sender/flight_state.h
@@ -1,10 +1,10 @@
 #ifndef FLIGHT_STATE_H
 #define FLIGHT_STATE_H
 
-extern int throttle_pct;
-extern int bank_angle;
-extern int pitch_angle;
+extern double throttle_pct;
+extern double bank_angle;
+extern double pitch_angle;
 extern double altitude;
-extern int rpm;
+extern double rpm;
 
 #endif // FLIGHT_STATE_H
